requirements.txt
torch
torchvision
streamlit
pillow
nltk
pandas
opencv-python
seaborn
tqdm
numpy 